moving base op new attribute strategies out of the templated headers
